+++ src/main/java/com/modagbul/BE/domain/fire/application/constant/FireFcmMessage.java
@@ -1,4 +1,4 @@
-package com.modagbul.BE.domain.fire.constant;
+package com.modagbul.BE.domain.fire.application.constant;
 
 import com.modagbul.BE.domain.mission.domain.entity.Mission;
 import lombok.Getter;
