+++ src/main/java/com/modagbul/BE/domain/fire/exception/FireAuthDeniedException.java
@@ -1,8 +1,6 @@
 package com.modagbul.BE.domain.fire.exception;
 
-import org.springframework.http.HttpStatus;
-
-import static com.modagbul.BE.domain.fire.constant.FireExceptionList.FIRE_AUTH_DENIED;
+import static com.modagbul.BE.domain.fire.application.constant.FireExceptionList.FIRE_AUTH_DENIED;
 
 public class FireAuthDeniedException extends FireException{
     public FireAuthDeniedException() {
