+++ src/main/java/com/modagbul/BE/domain/mission/presentation/MissionController.java
@@ -7,10 +7,10 @@ import com.modagbul.BE.domain.mission.application.dto.MissionRateDto;
 import com.modagbul.BE.domain.mission.dto.*;
 import com.modagbul.BE.domain.mission.domain.service.MissionBoardService;
 import com.modagbul.BE.domain.mission.domain.service.MissionService;
-import com.modagbul.BE.domain.usermission.constant.Status;
-import com.modagbul.BE.domain.usermission.dto.UserMissionDetailDto;
-import com.modagbul.BE.domain.usermission.dto.UserMissionStatusDto;
-import com.modagbul.BE.domain.usermission.service.UserMissionService;
+import com.modagbul.BE.domain.usermission.application.constant.Status;
+import com.modagbul.BE.domain.usermission.application.dto.UserMissionDetailDto;
+import com.modagbul.BE.domain.usermission.application.dto.UserMissionStatusDto;
+import com.modagbul.BE.domain.usermission.domain.service.UserMissionService;
 import com.modagbul.BE.global.dto.ResponseDto;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
