+++ src/main/java/com/modagbul/BE/domain/usermission/application/dto/UserMissionListDto.java
@@ -1,6 +1,6 @@
-package com.modagbul.BE.domain.usermission.dto;
+package com.modagbul.BE.domain.usermission.application.dto;
 
-import com.modagbul.BE.domain.usermission.constant.Status;
+import com.modagbul.BE.domain.usermission.application.constant.Status;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
