+++ src/main/java/com/modagbul/BE/domain/usermission/exception/NotFoundUserMissionsException.java
@@ -1,7 +1,7 @@
 package com.modagbul.BE.domain.usermission.exception;
 
 
-import static com.modagbul.BE.domain.usermission.constant.UserMissionExceptions.NOT_FOUND_USERMISSIONS;
+import static com.modagbul.BE.domain.usermission.application.constant.UserMissionExceptions.NOT_FOUND_USERMISSIONS;
 
 public class NotFoundUserMissionsException extends UserMissionException {
